Squashed commit of the following:
commit c8fb6c6788159b8ce826a0b08420ab4e385f547d
Merge: 9a13672 39137db
Author: Aiden2014 <54656326+Aiden2014@users.noreply.github.com>
Date:   Sun Apr 21 19:13:25 2024 +0800

    Merge pull request #10 from NJUPT-SAST/dev-Aiden

    fix: cru for /admin/judge

commit 39137db63afcc747044e9b54a0e3706f3d98dee6
Author: Aiden2014 <HiAidenLiu@gmail.com>
Date:   Sun Apr 21 19:11:58 2024 +0800

    fix: cru for /admin/judge

commit 9a136723decc3a40a9aeb972b0517acea6f04554
Merge: 10dabdb 24f9c1c
Author: Love98 <77888749+love98ooo@users.noreply.github.com>
Date:   Fri Apr 19 21:18:40 2024 +0800

    Merge pull request #9 from NJUPT-SAST/dev-love98

    fix: issues from ent update

commit 24f9c1cba702c0ad94f56b91dfbe0414e6dfb792
Author: Love98 <liaocai@sast.fun>
Date:   Fri Apr 19 20:20:17 2024 +0800

    fix: issues from ent update

commit 10dabdb489a249bc4e7dc819e4bb302833d5d9de
Merge: e2c894f e4b1af1
Author: Aiden2014 <54656326+Aiden2014@users.noreply.github.com>
Date:   Fri Apr 19 10:39:26 2024 +0800

    Merge pull request #8 from NJUPT-SAST/dev-Aiden

    Dev aiden

commit e4b1af16377dc6c7a26265cdd021d7c8cacad1a5
Author: Aiden2014 <HiAidenLiu@gmail.com>
Date:   Fri Apr 19 10:38:31 2024 +0800

    fix: conflict between current apis and ent schema

commit 0ea13fb8163363f5f5199f197f673f8af23130ef
Author: Aiden2014 <HiAidenLiu@gmail.com>
Date:   Thu Apr 18 20:53:01 2024 +0800

    feat: clear types of fields and name foreign keys

commit 2bc594bf4dbc5ebe062ece286aa3427e568b4a65
Merge: 85d9837 e2c894f
Author: Aiden2014 <HiAidenLiu@gmail.com>
Date:   Thu Apr 18 19:53:28 2024 +0800

    Merge branch 'main' into dev-Aiden

    # Conflicts:
    #	ent/client.go
    #	ent/loginsession.go
    #	ent/loginsession/loginsession.go
    #	ent/loginsession_create.go
    #	ent/loginsession_query.go
    #	ent/migrate/schema.go
    #	ent/schema/loginsession.go
    #	ent/schema/submit.go
    #	ent/submit.go
    #	ent/user/user.go
    #	ent/user/where.go
    #	ent/user_create.go
    #	ent/user_query.go
    #	ent/user_update.go
    #	openapi.yaml

commit 85d98375d6f3160dac6f640e4c67d4cbdb7fc232
Author: Aiden2014 <HiAidenLiu@gmail.com>
Date:   Thu Apr 18 19:49:49 2024 +0800

    refactor: fields and edges of ent schema

commit e2c894f6dffd323384ddc1f20749c2f36a5e9f68
Merge: 3367bc9 5d69b90
Author: Love98 <77888749+love98ooo@users.noreply.github.com>
Date:   Wed Apr 17 14:11:00 2024 +0800

    Merge pull request #7 from NJUPT-SAST/dev-love98

    feat: add user client apis

commit 5d69b90dd978e428ac0500d898b921ad06ac9324
Author: Love98 <liaocai@sast.fun>
Date:   Wed Apr 17 14:03:50 2024 +0800

    fix: add input for pretest

commit 4e930ff4efb2b62acd45e1a71c1a7673624228fb
Merge: e9ce572 3367bc9
Author: Love98 <liaocai@sast.fun>
Date:   Wed Apr 17 13:55:11 2024 +0800

    Merge branch 'main' into dev-love98

commit e9ce572e71586772f714087654aac0df61932cbe
Merge: 675cc4d 1a3d7b2
Author: Love98 <liaocai@sast.fun>
Date:   Wed Apr 17 13:47:22 2024 +0800

    Merge branch 'main' into dev-love98

commit 675cc4df91fa96bbe5db82f52bd6ca42253f6aed
Author: Love98 <liaocai@sast.fun>
Date:   Tue Apr 16 22:05:36 2024 +0800

    feat: add api GetSubmission

commit 8023c88aaa4bd80a571014a53b4d89c77e7701e2
Author: Love98 <liaocai@sast.fun>
Date:   Tue Apr 16 21:52:37 2024 +0800

    feat: add apis submit and pretest

commit 09589cf6046ab7ce8ac76af60cc89ef47dbc9250
Author: Love98 <liaocai@sast.fun>
Date:   Tue Apr 16 21:50:52 2024 +0800

    fix: ent schema error

commit 3367bc9be25cc7f7ef06790b1d7eaaa8bc87771c
Merge: 1a3d7b2 eebcd50
Author: Li Feng <maplestarplay@gmail.com>
Date:   Tue Apr 16 19:47:12 2024 +0800

    Merge pull request #6 from NJUPT-SAST/dev-lifeng

    Add Crud capability for user and group

commit e772d6d04ac1ec8227d6b23329decefa36055490
Author: Aiden2014 <HiAidenLiu@gmail.com>
Date:   Tue Apr 16 17:12:50 2024 +0800

    feat: add crud for admin/case

commit 5695476520e5b8fb464c5c581d2f97da8a5bb6f5
Author: Love98 <liaocai@sast.fun>
Date:   Mon Apr 15 20:00:04 2024 +0800

    feat: support rabbitmq

commit d469b2b9b6d9c40bb585359ad80ddcb865ad05ba
Author: Love98 <liaocai@sast.fun>
Date:   Sat Apr 13 22:33:46 2024 +0800

    feat: Get problems and problem

commit f34b25e68a2c8592af6c55e6c4bb144be7af8236
Author: Love98 <liaocai@sast.fun>
Date:   Tue Apr 2 18:46:02 2024 +0800

    fix: Makefile command error on Windows

commit 5a2c7ebc9e236d543efea87427ca74c1c63c9075
Author: Love98 <liaocai@sast.fun>
Date:   Tue Apr 2 18:45:13 2024 +0800

    feat: finish API join contest

commit 5de08878560c175fe687303ebc545a055a390335
Author: Love98 <liaocai@sast.fun>
Date:   Tue Apr 2 16:15:14 2024 +0800

    feat: support Redis and upgrade requirements

commit 63eb1acf5b8342076953a422064e89d8dae9024a
Author: Love98 <liaocai@sast.fun>
Date:   Tue Mar 26 22:40:06 2024 +0800

    feat: add user's APIs and implement Get Problems

commit 1de562451b1177aff7c9b1e59ffc6a9f25bc1d1c
Merge: 7eb2425 7e484d6
Author: Love98 <liaocai@sast.fun>
Date:   Tue Feb 20 18:27:54 2024 +0800

    Merge remote-tracking branch 'origin/dev-love98' into dev-love98

    # Conflicts:
    #	ent/client.go
    #	ent/loginsession.go
    #	ent/loginsession/loginsession.go
    #	ent/loginsession/where.go
    #	ent/loginsession_create.go
    #	ent/loginsession_query.go
    #	ent/loginsession_update.go
    #	ent/migrate/schema.go
    #	ent/mutation.go
    #	ent/schema/loginsession.go

commit 7eb24253d60384d67a86423b72bac42d1bc6cee1
Merge: d7df529 0d1e2bd
Author: Love98 <liaocai@sast.fun>
Date:   Tue Feb 20 17:29:54 2024 +0800

    Merge remote-tracking branch 'origin/dev-love98' into dev-love98

    # Conflicts:
    #	ent/client.go
    #	ent/contest.go
    #	ent/contest/contest.go
    #	ent/contest/where.go
    #	ent/contest_create.go
    #	ent/contest_query.go
    #	ent/contest_update.go
    #	ent/contestgroup.go
    #	ent/contestgroup/contestgroup.go
    #	ent/contestgroup/where.go
    #	ent/contestgroup_create.go
    #	ent/contestgroup_query.go
    #	ent/contestgroup_update.go
    #	ent/loginsession.go
    #	ent/loginsession/loginsession.go
    #	ent/loginsession/where.go
    #	ent/loginsession_create.go
    #	ent/loginsession_query.go
    #	ent/loginsession_update.go
    #	ent/migrate/schema.go
    #	ent/mutation.go
    #	ent/problem.go
    #	ent/problem/problem.go
    #	ent/problem/where.go
    #	ent/problem_create.go
    #	ent/problem_query.go
    #	ent/problem_update.go
    #	ent/problemcase.go
    #	ent/problemcase/problemcase.go
    #	ent/problemcase/where.go
    #	ent/problemcase_create.go
    #	ent/problemcase_query.go
    #	ent/problemcase_update.go
    #	ent/problemjudge.go
    #	ent/problemjudge/problemjudge.go
    #	ent/problemjudge/where.go
    #	ent/problemjudge_create.go
    #	ent/problemjudge_query.go
    #	ent/problemjudge_update.go
    #	ent/schema/contest.go
    #	ent/schema/contestgroup.go
    #	ent/schema/loginsession.go
    #	ent/schema/problem.go
    #	ent/schema/problemcase.go
    #	ent/schema/problemjudge.go
    #	ent/schema/submit.go
    #	ent/schema/submitcase.go
    #	ent/schema/submitjudge.go
    #	ent/submit.go
    #	ent/submit/submit.go
    #	ent/submit/where.go
    #	ent/submit_create.go
    #	ent/submit_query.go
    #	ent/submit_update.go
    #	ent/submitcase.go
    #	ent/submitcase/submitcase.go
    #	ent/submitcase/where.go
    #	ent/submitcase_create.go
    #	ent/submitcase_query.go
    #	ent/submitcase_update.go
    #	ent/submitjudge.go
    #	ent/submitjudge/submitjudge.go
    #	ent/submitjudge/where.go
    #	ent/submitjudge_create.go
    #	ent/submitjudge_query.go
    #	ent/submitjudge_update.go
    #	ent/user.go

+++ app/user/contest/internal/biz/problem.go
@@ -9,14 +9,14 @@ type Problem struct {
 	ID      int64
 	Title   string
 	Content string
-	Point   int32
-	Index   int32
+	Point   int16
+	Index   int16
 }
 
 type ProblemRepo interface {
 	ListProblem(ctx context.Context, contestID int64) ([]*Problem, error)
 	GetProblem(ctx context.Context, problemID, contestID int64) (*Problem, error)
-	GetProblemCaseVer(ctx context.Context, problemId int64) (int32, error)
+	GetProblemCaseVer(ctx context.Context, problemId int64) (int8, error)
 }
 
 type ProblemUsecase struct {
@@ -44,6 +44,6 @@ func (uc *ProblemUsecase) GetProblem(ctx context.Context, contestID int64, probl
 	return problem, nil
 }
 
-func (uc *ProblemUsecase) GetProblemCaseVer(ctx context.Context, problemId int64) (int32, error) {
+func (uc *ProblemUsecase) GetProblemCaseVer(ctx context.Context, problemId int64) (int8, error) {
 	return uc.repo.GetProblemCaseVer(ctx, problemId)
 }
